+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changes
 dev (master)
 ++++++++++++
 
-* ...
+* Improved url parsing in ``urllib3.util.parse_url`` (properly parse '@' in
+  username, and blank ports like 'hostname:').
 
 
 1.7.1 (2013-09-25)